mcp-servers/requirements.txt
mcp>=1.0.0
google-search-results>=2.4.2
python-dotenv>=1.0.0
